Test client_keyblock kdcpreauth callback
Add internal clpreauth and kdcpreauth modules named "test" which can
exercise the client_keyblock callback (as well as get_string and
get_as_key on the client side).  Add tests to t_etype_info.py to
verify that the callback matches the etype info sent by the KDC.

In the KDC's load_preauth_plugins(), correct a test for the end of
pa_type_list so that we can use a negative preauth type number for the
test module.  (RFC 4120 reserves negative preauth type values for
unregistered use.)

ticket: 8200