In kpropd, poll after finishing resync
When kpropd operates in iprop mode, full resyncs are handled by a
child process.  After a full resync, we want to poll for incremental
updates, as the dump we received may have come from a pre-existing
dump file which was not current.  To make this polling happen
promptly, signal the parent process from the child process after a
dump is received.

With this change, t_iprop.py no longer has to prod kpropd after a full
resync occurs, so remove that logic.

ticket: 8161