Document kpropd -t and fix it in iprop mode
If kpropd is asked to run just once, don't exit after starting a full
resync; we want to wait for the fullprop child to process the request,
and then request incremental updates afterwards.  Also don't exit from
do_standalone() in the fullprop child, in case multiple full resyncs
are required to get the database up to date.

Document the -t flag in kpropd.rst.

ticket: 8161