gssapi: Implement gss_krb5_no_ci_flags()
BUG: http://krbdev.mit.edu/rt/Ticket/Display.html?id=6938

GSS-SPNEGO which is a non-standard SASL mechanism omits the usual wrap
exchange after the GSS context is established, As a result, it does not
support authzids, does not negotiate a maximum message size, and
implicitly negotiates a security layer based on the GSS flags asserted
by the client.  There does not appear to be any way for the server to
disclaim support for a security layer, so if the client asserts GSS
flags the server doesn't want to support, the server has no alternative
but to reject the connection.

This allows to NOT force confiendality and integrety flags on the
client.

Signed-off-by: Andreas Schneider <asn@samba.org>