Add leash column for the ccache name
Previously, it was fairly easy to set the default cache to MSLSA:,
which is not collection-enabled (as opposed to the normal default
of API:, which is), and the non-collection behavior proved confusing
to many users.

Ideally there would be an option to choose the output ccache in
the "Get Ticket" window, but that is more complicated to implement
than just a display of what cache a given ticket is in.  This
extra display column should still help to alleviate user confusion.

ticket: 8207 (new)
queue: kfw
tags: pullup
target_version: 1.13.3