Make ribbon home tab more pretty
In the absence of a label title for a given command group, the
text "MS Shell Dlg" was inserted, which is somewhat confusing to
the user.  Give the ticket-maniuplation group a title consisting
of a space, which removes the confusing text from the display.

(cherry picked from commit bfba2d235370a3902faeeaad8a54d8a2dcc3427b)

ticket: 8206
version_fixed: 1.13.3
status: resolved