Put focus on password field when principal is set
In the Leash "Get Ticket" window, under some user-customized
configurations, the principal field can be pre-populated.  In this case,
it makes sense for the default focus to be on the password field,
since most users will only be using a single principal and should
go directly to typing their password.

The focus was already set to the password field in the case when
the principal was not modifiable (such as when the "Get Ticket" dialog
was opened by an application requesting a specific ticket).

ticket: 8032
tags: pullup
target_version: 1.13.3