Don't assert GSS_C_MA_NOT_DFLT_MECH in mechglue
gss_inquire_attrs_for_mech() should not add GSS_C_MA_NOT_DFLT_MECH to
mechs which aren't the default.  The attribute means "MUST NOT be used
as a default mechanism" (RFC 5587) and is intended to be used by the
mechglue.  It does not mean "is not the default mech".

ticket: 8021