Pick MFC version based on VisualStudioVersion
Rather than hardcode support for VS2010 and VS2012, assume that the
right MFC version can be obtained by removing the "." in the
VisualStudioVersion variable.  This is true for VS2010, VS2012,
VS2013, and VS2015 at least.  If VisualStudioVersion it is not set,
fall back to using MFC100(D) as before.

ticket: 8231