Remove windows/gina
It has only received cosmetic and build fix changes since it was
introduced in 1997 and doesn't do anything useful.  The motivation for
removal at this point is to avoid spending time on build fixes for
newer tool chains like VS2015, and because GINA modules are only
supported by Windows versions prior to Windows Vista.

https://support.microsoft.com/en-us/kb/925520

ticket: 8231