Resolve krb5 GSS creds if time_rec is requested
The code normally tries to defer credential acquisition to a later
time.  However, if the application requests the lifetime, the code
needs to resolve the credential and return the actual expiration time.
Returning 0 would cause the application to think credentials are
expired.

In the mechglue, pass through null time_rec pointers to the mech so
that the mech knows whether it was requested.  In SPNEGO, pass through
time_rec to the mech when acquiring creds, via a new parameter to
get_available_mechs().

[ghudson@mit.edu: minor style changes; edit and expand commit message]

ticket: 8235 (new)