Adjust SPNEGO Microsoft krb5 OID tests
In the Microsoft krb5 OID tests in t_spnego.c, use the proper krb5 OID
for the framing of the mech token, which better reflects the behavior
of Microsoft clients that use the wrong OID.  Move the test to a
separate function and run it twice, once with an acceptor cred and
once without.

Also add a comment noting that the reselection test no longer
exercises what it was designed to.  We can't really fix that until we
add gss_acquire_cred_with_cred() or similar.