Update SPNEGO hintName value to current spec
[MS-SPNG] currently specifies that the hintName field of NegHints
should contain "not_defined_in_RFC4178@please_ignore".  Heimdal
implements this behavior, but we instead try to include a display
name.  Implement the currently specified behavior, and add a test to
t_spnego.c to verify that the expected hint token is generated.

Further cleanup is possible; the negHints encoding is now constant (so
it does not need to be generated dynamically), and we could avoid
abusing the mechListMIC parameter of make_spnego_tokenInit_msg() to
transport it.

ticket: 8236 (new)