Add tabular dump capability to kdb5_util
This new kdb5_util tabdump command provides a reporting-friendly
tabular dump format for the KDC database.  This format is also
suitable for importing into relational databases for queries.  Output
is in tab-separated or CSV format.  The user can select an output
table with a fixed number of columns.

Currently, this only provides tables for a subset of the available
principal data.  This includes making visible some data that is hidden
in hexadecimal strings in the tl_data of the ordinary dump format.

ticket: 8243