Renaming reconf script to align more closely with autotools conventions
There are certain conventions users expect when working with a project
that uses autotools. Deviating from those conventions adds confusion and
complexity to a project, and let's face it, krb5 is not exactly a
shining example of a convention-conforming project. In an effort to try
to simplify the krb5 build, I am moving and renaming the src/util/reconf
script to src/bootstrap. The bootstrap name was chosen since the
autotools documentation says it is the preferred name. For more
information on this please see:

https://goo.gl/JRvpYK
https://goo.gl/V1i521

ticket: 8299