Add IPv6 address logging support to audit plugin
The jsonwalker.py test was failing due to the audit plugin only logging
the ip for IPv4. The audit plugin should log the ip address for both
IPv4 and IPv6. This commit renames the json "ipv4" field in the
"fromaddr" json object to "ip" and sets the field to both IPv4 and IPv6
addresses. Any parsers of the audit log can tell which address is stored
in the "ip" field by checking the "type" field in "fromaddr". The
jsonwalker.py reference json, au_dict.json, is also updated to handle
this new field name.

old audit:

{"fromaddr":{"type":<int>,"length":<int>,"ipv4":[<4 int>]}

where "ipv4" only occurs when type is ADDRTYPE_INET

new audit:

{"fromaddr":{"type":<int>,"length":<int>,"ip":[<<length> int>]}

where "ip" only occurs when type is ADDRTYPE_INET or ADDRTYPE_INET6

ticket: 8298