Cast away constness when assigning to krb5_data
Some password-changing library functions take a const char * parameter
but try to assign it to krb5_data.data, which isn't const.  PR #364
causes some compilers to produce errors in such situations, so cast
away the constness.  This is almost certainly safe because of the
nature of the code that consumes these krb5_data values.