Add the ability to lock down principal keys
A new attribute named KRB5_KDC_LOCKDOWN_KEYS can be set on principals.
This flag prevents keys for the principal from being extracted or set
to a known value by the kadmin protocol.  Principals with this flag
cannot be deleted or renamed, and cannot have keys set by setkey or
chpass.  chrand operations are allowed, but keys are not returned.
This attribute can be set via the modify operation but cannot be
reset; an authorization error is resturned if an attempt to reset it
is performed.

When creating a KDB, set the lockdown flag on the krbtgt and kadmin
principals.

[ghudson@mit.edu: squash with t_kadmin_acl.py commit; condense commit
message]

ticket: 8365 (new)