Add the ability to lock down principal keys.
A new attribute named KRB5_KDC_LOCKDOWN_KEYS can be set on principals.
This will indicate that keys for this principal can never be extracted
via the kadmin protocol.
Any operation that would cause potential disclosure of keys for the
named principal is disallowed when this attribiute is set and an
authorization error is returned.

When this attribute is set, the principal cannot be deleted or renamed,
keys cannot be set, all chpass operations are denied while chrand
operations are allowed, but in this case keys are note returned.

This attribute can be set via the modify operation but cannot be reset,
an autorization error is resturned if an attempt to reset it is performed.

Signed-off-by: Simo Sorce <simo@redhat.com>