Add Github provided OS and IDE files to .gitignore
It is a good idea to have a fairly complete .gitignore file.  It is
helpful for devs to know just what shouldn't end up in the repo, as
well as preventing accidental commits and even random trash ending up
in master.

Many of the files here added from the Github gitignore project
found here:

https://github.com/github/gitignore