Only store latest keys in key history entry.
A bug was found in the kadm5 server where if a password is changed
with the -keepold option the history entry that is created once the
password is changed again contains both the latest password and the
one that was kept.  Fix create_history_entry to only store the
latest key (largest kvno) in the history entry.  Also, add a test
to ensure that the bug is fixed.

More information may be found here:

http://mailman.mit.edu/pipermail/krbdev/2014-July/012084.html

ticket: 8354