Only store latest keys in key history entry
If a password is changed with the -keepold option, then changed again,
the history entry contains both the latest password and the one that
was kept.  Fix create_history_entry to only store the latest kvno in
the history entry.  Also add a test to ensure that the bug is fixed.

ticket: 8354