Implement password history in LDAP plugin.
The password history is stored in the kerberos LDAP schema attribute
'krbPwdHistory', with one history entry per attribute.  When the
history is decoded the history entries are sorted by kvno in
descending order, with the next replacement key set to the last
key entry, or the key history entry with the smallest kvno.

This commit is based on code commited by
Tomas Kuthan <tkuthan@gmail.com>.

ticket: 5889