Implement password history in LDAP KDB module
The password history is stored in the kerberos LDAP schema attribute
'krbPwdHistory', with one history entry per attribute.  When the
history is decoded, the history entries are sorted by kvno with the
next replacement key set to the end of the list.  Based on a patch
from Tomas Kuthan.

ticket: 5889