Add hints for -A flag to kdestroy
When using a collection ccache, a user accustomed to the FILE ccache behavior
may not be aware of all active caches and the default kdestroy command
could make it seem like there is no active cache left. Print a warning
to use -A after kdestroy if there are other caches.