Report inquire_attrs_for_mech mechanism-related failures
Previously, gss_inquire_attrs_for_mech would return a list of mech
attributes that it knew about when given a bad mech oid or a mechanism
which dod not provide a gss_inquire_attrs_for_mech() method.  It seems
more useful to just report the failure to the application rather than
allowing it to continue with a faulty mechanism.