Default to LSA when TGT in LSA is inaccessible
When UAC is enabled and a domain user with Administrator privileges
logs in the TGT is inaccessible.  This is because if the TGT is
accessible it may be allow a non-elevated user to bypass the UAC.
This causes a bug where the ms2mit.exe copies the current tickets in
the LSA cache to the API ccache except the TGT, which prevents a user
session from getting new tickets.  In this case, ms2mit will now set
the LSA cache to the default cache to workaround this issue.

[tlyu@mit.edu: edit commit message]

ticket: 8390