Allow "includedir" to accept config files ending with ".conf"
Since the main config file is krb5.conf, the most obvious thing to name
a config file discovered through the "includedir" directive is
"REALM.COM.conf".

Once a user does that, they spend the next few minutes yanking their
hair out trying to figure out what syntax error they've got that causes
kinit not to find their realm.  Eventually, the user strace's kinit and
discovers it's ignoring the file altogether, but it's reading another
file next to it in the directory.  At this point a file renaming game
ensues, and the user discovers ".conf" is somehow not allowed at the end
of filenames.

If the user is unlucky, as I was, they named their config file
"redhat.conf" and not "redhat.com.conf", and so never figured out it's
just that all dots are banned until quite a bit of debugging later,
instead looking for something like an inverted strcmp() looking for
.conf itself.

This patch enables config files that end with ".conf".

Signed-off-by: Peter Jones <pjones@redhat.com>