Remove logger.c compile-time syslog conditionals
In logger.c, remove conditionals around uses of syslog.h, syslog(),
openlog(), closelog(), and syslog priority constants.  This header and
these symbols are used unconditionally in other non-Windows parts of
the tree.  Leave the conditionals around facility constants for now,
as not all of the facility constants are specified in POSIX.