Add kprop and iprop default_realm tests
Add tests to t_iprop.py and t_kprop.py that exercise cases where
default_domain and domain_realm maps differ, as well as overriding the
default_realm with the -r argument.  This includes the testing of -r
with kadmind, and an update of expected ulog numbers for tests following
the addition of an incremental test.  Also refactor some common code in
t_kprop.py to use in the new tests.